refactor graph build into module
